for tests, set time zone to Europe/London
This enables reliable testing of time zone behavior. For example, if
we'd like to assert that a control shows local time, we can mock the
time so that it is not equal to UTC, and make an assertion for the
numbers displayed on the screen. With this setting, we can choose such a
date knowing that the local time zone is Europe/London. Such tests would
have also worked without this setting, but only when executed on a
machine with the proper time zone settings. This ensures that these
tests run consistently across machines in different time zones.

What would be even better if the time zone could be set from individual
tests, which I think is also possible but would take more work.