hide `Test Sections` when there aren't any
This could affect the `/view` and `/attempt` pages as well, but I think
each quiz has at least one section. Even if there were tests without
sections, it'd make no sense to show a header and then not list any
sections. The `QuizView` component always passes an empty array for
sections, so the outcome is that sections are hidden from that component
only.