make sure errors are still shown
Previously, the synchronous operations after firing the requests would
finish before the requests were actually sent. This meant it was okay to
enable error reporting after triggering the request. Now that we wait
for the requests to finish, it's important to enable error reporting
before firing the requests.

It doesn't really matter whether we reset the feedback field before or
after the requests were sent, as the box only appears after the last
request finishes, but it's still safer to do before sending the requests.