Attempt some test fixes
The modal used for the SetAssignments component is no longer in the
component but the global ActiveModal. So we must render the whole app
and not one component to test this.
If we render the whole app, setting an initial path does nothing, which
breaks some of the tests.
Another test was broken because we no longer leave the Set Assignment
modal open after setting an assignment.