Fix remaining failing tests for Set Assignments
This does two things; no longer tests that the URL hash anchor works,
but instead just opens the modal to test the behaviour. It is not clear
that testing the modal hash anchor behaviour can reliably be tested
outside of a real browser.
It also fixes the final two tests which were reliant on the same render
function as the ones fixed in the previous commit.