Restore "passed=correct" in assign progress totals
We used to count "passed" questions as complete and correct in the
totals used on the markbook. We no longer do this on redesign-2024, but
do still on master. For now, replicate the existing functionality (which
the previous commit had accidentally removed) until we have more clarity
here.
Also rename "tickCount" to the more obvious "correctQuestionPagesCount"
since that is what it is.