Fix group progress page quiz download link
This moves all the download link generation functions into progress.ts,
so that there is one source of truth. It then fixes the issue in
AssignmentProgressGroup where the assignment link was used regardless of
whether it was a quiz or not.

The rest of the changes are just optimising imports and fixing poor
whitespace formatting.