Fix sorting by total correct on Assignment Progress
This makes the same change as 946b0b30f4f486e7b8efbd599ec3fc73960bc852
to rename "tickCount" to the clearer "correctQuestionPagesCount".

However, the bug the previous commit (54c38d0e4e08bd1b7c6cb09fb3626331abec52f2)
was attempting to fix only applied to quizzes, which don't have pages but
also had completely meaningless value for "tickCount" computed in
QuizTeacherFeedback. This commit moves to using the part count for
quiz "page" totals to mean the sorting can be done consistently.

It also fixes `questionMarks?.[q?.id ?? 0]` which could use a real but
irrelevant value if the question ID was not found in the marks list.
Using -1 instead of zero ensures that the nullish coalescing would fall
though in that case instead of using the first part's result.