Add new Drag and Drop question type
A new question type to supersede and improve upon Cloze questions, this reuses most of the Cloze logic while replacing the index-based list approach of submitting answers.

The format of Cloze questions' answers is one list of entries, where the index into the list represents which drop zone an answer is in, and the items in the list are `ItemDTO`s containing an `id` and `value` field. The new approach is to give all drop zones a unique ID, which can simply be the index, or can be more advanced (useful for e.g. figure numbering), which is also stored alongside the (item) `id` and `value` inside a new extension of `ItemDTO`, `DndItemDTO`.