Restructure routes to allow group name info on assignment view
The new approach means all group *name* information (which is only currently accessible via the endpoint that loads all groups...) is loaded exactly once on the loading of the markbook, and is retained and never once requested again until leaving the markbook. Assignment details and group *member* information are still requested on-demand, as we do not want to preload all of that when it is likely not needed. Hoorah!