Restore padding for other uses of AdaCard
I don't like this way at all, but we use it in a lot of places so not the easiest to remove. In short, spacing between multiple cards ought to be entirely managed by the parent, preferably through `gap`, but `row row-cols-x` also puts padding on elements inside (but then not vertically??); in any case, it certainly should not be set on the card itself, as this reduces reusability.