Add 2nd ed. maths book to My Progress, remove 1st
We only want the latest books showing for users on My Progress unless
they have already answered questions from that book.

Remove an unnecessary extra list of book titles that was inconsistent
with the others. This requires adding physics_skills_14 to the list of
books. The only side effect of this is that the 1st Edition Pre-Uni
Maths book no longer shows on the Set Assignments books modal, so that
we can hide the unwanted physics book.