focus input fields when 2fa code is requested
Anton has asked for this. There is an eslint rule against using the
autofocus attribute. After a web search, the concerns appear to be:
- on mobile, autofocus automatically opens the keyboard
- on large pages, autofocus can cause the page to scroll unexpectedly
- on a screen reader, it interrupts the narration by forcing the reader
  on the field

I've checked on my phone (Android, Chrome), and the keyboard does not
open automatically. The page is small, so no scrolling occurs. I haven't
tested with a screen reader, but that field is the only field on the
page, so even for a screen reader user, jumping there seems like
the right thing to do. This UI is only shown for Isaac staff.