Overhaul all old-style font management
All text should now be styled with one of the three font families (`$primary-font`, `$secondary-font`, `$tertiary-font`). Note that for Ada, these are all the same anyway. To achieve a bold / italic font, rather than setting this in the font *family*, adjust the relevant field (`font-weight` or `font-style`, respectively).

This also removes all of the old individual font files and replaces them with the modern variable weight ones. This is what we do for Hanken Grotesk and it does not seem to have caused any problems since the redesign, so I am confident this should be safe.