Improve issue with tooltip (+arrow) overlapping focus area
This bug causes the tooltip to rapidly appear / disappear as the mouse is detected as being over / not over the tooltip focus area. For a reason I can't work out, it seems the tooltip arrow is being excluded from the popper positioning calculation, so it overlaps commonly. This change adds a `pointer-events: none` to the arrow, so that it can't cause this issue. I would add this to the whole tooltip, but this behaviour is not wanted on `autohide={false}` tooltips, so this is a quick compromise for now.