Use recursive height calculations on nested collapsibles
Restructures the implementation of collapsibles to allow a second method through which height can be recalculated: if a child recalculates its height, its parent should too. This more naturally follows the logical structure of what is happening and avoids reliance on rerenders to update parents.

The old method of requiring a `useEffect` dependency on `children` is still required, as non-collapsible components updating may reflow the height and thus should cause a recalculation (c.f. Physics QF topic selector). The new method specifically permits updates when a child's `data-targetheight` prop changes, which the `useEffect` approach could not deal with as this is not stateful.