Show correct aria label on clear date button
`<Button>`s with the `close` prop have a default aria label of "close" which is not overridden by putting a different aria label on the component itself. Therefore to have a sensible aria label on this button we can't use `close`. If we instead make this a `<button>` we can use `btn-close` to keep the same styling without setting unwanted properties.