use `@stylistic/indent` rule instead of `indent`
According to the [docs](https://eslint.org/docs/latest/rules/indent),
`indent` has been deprecated and this is the recommended replacement.

The 2025 September package updates, merged in PR1750 also broke the
rule, so now `yarn run lint` fails after a recursive call reaches max
stack depth. Since the rule is deprecated, eslint might or might not
fix it, so I thought it was easier to just move to the maintained
version.

My personal preference would be to use prettier for formatting instead
of `eslint-stylistic`, but I went with what I thought was easier for
now.