remove book page for 2nd edition physics book
The current edition is the 3rd edition. URLs for the 2nd edition book
page now redirect to the 3rd edition book page. Lisa wanted as
few people as possible to keep using the old questions, and this is part
of our attempt to steer people away from the old edition.