don't break build when a suppression is fixed
We've noticed that after somebody fixes a suppressed lint problem, the
build breaks until the suppressions file is pruned. Jaycie and others
have found that this was annoying, as it means the build effectively
discourages fixing lint problems.

We've decided the best way forward will be a precommit hook that
automatically prunes suppressions for fixed problems. Until we build
that, this change allows the build to pass even when the suppressions
file is out of date.