hide irrelevant fields, topics from GCSE Maths QF
This is implemented with a potential extension in mind. If, in the
future, we'd like to hide more fields or topics, depending on whether
there are any available questions for the context, we can just change
the backend to return the list of available tags for the context, and
change getAllowedTags to return these.