Improve QR lazy import error on connection loss
We ought to have a default for all lazy imports, but in all other cases I could find, the lazy import is either used immediately on page load or never during the lifecycle of the page. This QR work was an exception as this could be loaded while on an active page, without a page transition, owing to the panel loading on a button press. As such, for all other cases the browser's "No internet" will show instead.