Replace all Route-style components with plain Routes
RR6+ only allows <Route/> and <React.Fragment/> as children of a <Routes/> block, even if the component returns one of these. As such, I have moved the tracking part of tracked routes into onPageLoad, the figure numbering into a parent of the router (TODO: test this works), and replaced StaticPageRoutes and ExternalRedirects with <Route/>-based equivalents.